backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi[standard]==0.109.0
uvicorn==0.27.0
pytest==8.0.0
httpx>=0.28.1
black==24.1.1
flake8==7.0.0
SQLAlchemy>=2.0.28
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
passlib>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
google-genai>=1.0.0
langchain-google-genai>=4.0.0
tenacity>=8.2.0
faiss-cpu==1.9.0
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2
biopython==1.83
requests==2.31.0
sentence-transformers==2.5.1
transformers==4.38.2
torch==2.2.1
rank-bm25==0.2.2
spacy==3.7.4
scikit-learn==1.4.1.post1

